tools: Move io utils to cpp (#5869)
In preparation for more code in that file, which would get unwieldy in a
header.
