+++ BUILD.gn
@@ -1466,6 +1466,15 @@ source_set("spvtools_tools_util") {
   configs += [ ":spvtools_internal_config" ]
 }
 
+source_set("spvtools_tools_io") {
+  sources = [
+    "tools/io.cpp",
+    "tools/io.h",
+  ]
+  deps = [ ":spvtools_headers" ]
+  configs += [ ":spvtools_internal_config" ]
+}
+
 if (spvtools_build_executables) {
   executable("spirv-as") {
     sources = [ "tools/as/as.cpp" ]
@@ -1473,6 +1482,7 @@ if (spvtools_build_executables) {
       ":spvtools",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
     ]
     configs += [ ":spvtools_internal_config" ]
   }
@@ -1483,6 +1493,7 @@ if (spvtools_build_executables) {
       ":spvtools",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
     ]
     configs += [ ":spvtools_internal_config" ]
   }
@@ -1493,6 +1504,7 @@ if (spvtools_build_executables) {
       ":spvtools",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
       ":spvtools_val",
     ]
     configs += [ ":spvtools_internal_config" ]
@@ -1508,6 +1520,7 @@ if (spvtools_build_executables) {
       ":spvtools",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
     ]
     configs += [ ":spvtools_internal_config" ]
   }
@@ -1519,6 +1532,7 @@ if (spvtools_build_executables) {
       ":spvtools_opt",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
       ":spvtools_val",
     ]
     configs += [ ":spvtools_internal_config" ]
@@ -1532,6 +1546,7 @@ if (spvtools_build_executables) {
       ":spvtools_opt",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
       ":spvtools_val",
     ]
     configs += [ ":spvtools_internal_config" ]
@@ -1552,6 +1567,7 @@ if (!is_ios && !spirv_is_winuwp && build_with_chromium && spvtools_build_executa
       ":spvtools_reduce",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
       ":spvtools_val",
       "//third_party/protobuf:protobuf_full",
     ]
@@ -1571,6 +1587,7 @@ if (!is_ios && !spirv_is_winuwp && spvtools_build_executables) {
       ":spvtools_reduce",
       ":spvtools_software_version",
       ":spvtools_tools_util",
+      ":spvtools_tools_io",
       ":spvtools_val",
     ]
     configs += [ ":spvtools_internal_config" ]
