+++ test/reduce/CMakeLists.txt
@@ -31,6 +31,7 @@ add_spvtools_unittest(TARGET reduce
         structured_construct_to_block_test.cpp
         structured_loop_to_selection_test.cpp
         validation_during_reduction_test.cpp
+        ${spirv-tools_SOURCE_DIR}/tools/io.cpp
         LIBS SPIRV-Tools-reduce
         )
